moved back the routes to its place
